create top nav and display it to small screens only
